Fixes to space plugin
