+++ packages/data-access-plugin/package.json
@@ -28,8 +28,8 @@
         "build": "tsc --build --pretty",
         "build:watch": "yarn build --watch",
         "test": "jest",
-        "test:watch": "jest --coverage=false --notify --watch --onlyChanged",
-        "test:debug": "env DEBUG=\"${DEBUG:-*teraslice*}\" jest --detectOpenHandles --coverage=false --runInBand"
+        "test:watch": "jest --coverage=false --notify --watch --onlyChanged --forceExit",
+        "test:debug": "env DEBUG=\"${DEBUG:-*teraslice*}\" jest --detectOpenHandles --coverage=false --runInBand --forceExit"
     },
     "dependencies": {
         "@terascope/data-access": "^0.8.0",
