+++ packages/data-access-plugin/src/spaces/dynamic-server.ts
@@ -1,13 +1,9 @@
-
 import { get, TSError } from '@terascope/utils';
 import * as apollo from 'apollo-server-express';
 import accepts from 'accepts';
 import { json } from 'body-parser';
 import { User } from '@terascope/data-access';
-import {
-  renderPlaygroundPage,
-  RenderPageOptions as PlaygroundRenderPageOptions,
-} from '@apollographql/graphql-playground-html';
+import { renderPlaygroundPage, RenderPageOptions as PlaygroundRenderPageOptions } from '@apollographql/graphql-playground-html';
 /* Don't think it's exported normally, get directly */
 import { graphqlExpress } from 'apollo-server-express/dist/expressApollo';
 
@@ -16,10 +12,7 @@ import getSchemaByRole from './dynamic-schema';
 import { makeErrorHandler } from '../utils';
 
 export class DynamicApolloServer extends apollo.ApolloServer {
-    applyMiddleware({
-        app,
-        path = '/graphql',
-    }: apollo.ServerRegistration) {
+    applyMiddleware({ app, path = '/graphql' }: apollo.ServerRegistration) {
         /* Adds project specific middleware inside, just to keep in one place */
         // @ts-ignore
         const loginErrorHanlder = makeErrorHandler('Failure to login user', this.logger);
@@ -29,13 +22,13 @@ export class DynamicApolloServer extends apollo.ApolloServer {
         app.use(path, json(), async (req, res, next) => {
             // @ts-ignore
             loginErrorHanlder(req, res, async () => {
-                const user = await utils.login(get(req, 'aclManager'), req) as User;
+                const user = (await utils.login(get(req, 'aclManager'), req)) as User;
                 if (user.role == null) {
                     throw new TSError(`User ${user.username} missing role `, {
                         statusCode: 403,
                         context: {
                             user,
-                        }
+                        },
                     });
                 }
 
@@ -45,10 +38,7 @@ export class DynamicApolloServer extends apollo.ApolloServer {
                     // but I don't think it needs any overriding
                     const accept = accepts(req);
                     const types = accept.types() as string[];
-                    const prefersHTML =
-                    types.find(
-                        (x: string) => x === 'text/html' || x === 'application/json',
-                    ) === 'text/html';
+                    const prefersHTML = types.find((x: string) => x === 'text/html' || x === 'application/json') === 'text/html';
 
                     if (prefersHTML) {
                         const playgroundRenderPageOptions: PlaygroundRenderPageOptions = {
@@ -68,8 +58,8 @@ export class DynamicApolloServer extends apollo.ApolloServer {
                 const { schema, ...serverObj } = this;
 
                 schemaErrorHanlder(req, res, async () => {
-                        // TODO: check if user is false
-                        // @ts-ignore
+                    // TODO: check if user is false
+                    // @ts-ignore
                     const roleSchema = await getSchemaByRole(req.aclManager, user, this.logger, this.pluginContext);
 
                     /**
@@ -91,7 +81,7 @@ export class DynamicApolloServer extends apollo.ApolloServer {
                             playground: {
                                 settings: {
                                     'request.credentials': 'include',
-                                }
+                                },
                             } as apollo.PlaygroundConfig,
                         })
                     )(req, res, next);
